added docs to some often used events
