+++ src/main/java/at/hannibal2/skyhanni/events/PlaySoundEvent.kt
@@ -7,6 +7,19 @@ import at.hannibal2.skyhanni.utils.NumberUtil.roundTo
 import at.hannibal2.skyhanni.utils.SoundUtils
 import at.hannibal2.skyhanni.utils.SoundUtils.playSound
 
+/**
+ * Fired when the client is about to play a sound.
+ *
+ * Fired on the main client thread via a Mixin into `SoundEngine.play`.
+ *
+ * This event is cancellable. Cancelling it prevents the sound from being played.
+ * Use [replaceWithOther] to cancel the current sound and play a different one instead.
+ *
+ * @param soundName the Minecraft sound identifier, without the `minecraft:` namespace prefix
+ * @param location the world position where the sound originates
+ * @param pitch the pitch of the sound
+ * @param volume the volume of the sound
+ */
 @PrimaryFunction("onPlaySound")
 class PlaySoundEvent(
     val soundName: String,
